Extract kernel initialization code into a module
